feat(cli/update): instead of detecting package manager and global/local install, inquire it
